pyproject.toml
[tool.poetry]
name = "font-back"
version = "0.1.0"
description = ""
authors = ["zeroswan <zeroswan@outlook.com>"]

[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true 

[tool.poetry.dependencies]
python = ">=3.6"
django = "^3.0.4"
djangorestframework = "^3.11.0"
mongoengine = "^0.19.1"
django-rest-framework-mongoengine = "^3.4.1"
django-filter = "^2.2.0"
Pillow = "^7.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
isort = "^4.3.21"
black = "^19.10b0"
flake8 = "^3.7.9"

[tool.black]
line-length=100
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

